update_ttl.py: initial commit
Scan a given table and look to rows that are either too old
or don't have a TTL set given the specific TTL value.

If too old row is detected - delete it at the spot.
Otherwise of the row doesn't have a TTL overwrite it using the TTL
value that would correspond to the given TTL if it would have been
used during the row write time.

TODO:
  * Allow providing DELETE and INSERT ops CLs. Currently we use ALL and
LOCAL_QUORUM correspondingly.
  * Improve the script style in general: see FIXME and TODOs in the
code

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>